Calling current weather api fixed.

+++ app/src/main/java/pl/pgalinski/openwatherapicompose/model/current/weather/CurrentWeather.kt
@@ -53,8 +53,8 @@ data class Fall(
 
 @Parcelize
 data class Coord(
-    val lon: Long,
-    val lat: Long
+    val lon: Float,
+    val lat: Float
 ) : Parcelable
 
 @Parcelize
